+++ src/mightybug_a/pid.h
@@ -7,7 +7,7 @@
 #define MAX_PID_ERROR 2000
 #define PID_CONTROL_DIVISOR 100
 
-int32_t pid(int32_t error);
+int16_t pid(int16_t error);
 void reset_pid();
 
 void set_kp(int);
